+++ src/Dictionary.php
@@ -102,7 +102,7 @@ final class Dictionary implements StructuredFieldOrderedMap
         return [] === $this->members;
     }
 
-    public function isNotEmpty(): bool
+    public function hasMembers(): bool
     {
         return !$this->isEmpty();
     }
@@ -148,15 +148,15 @@ final class Dictionary implements StructuredFieldOrderedMap
     /**
      * Returns all containers Item values.
      *
-     * @return array<string, InnerList<int, Item>|Token|ByteSequence|float|int|bool|string>
+     * @return array<string, InnerList<int, Item>|float|int|bool|string>
      */
     public function values(): array
     {
-        $mapper = function (Item|InnerList $item): InnerList|Token|ByteSequence|float|int|bool|string|null {
+        $mapper = function (Item|InnerList $item): InnerList|float|int|bool|string|null {
             try {
                 $member = self::filterForbiddenState($item);
 
-                return $member instanceof Item ? $member->value : $member;
+                return $member instanceof Item ? $member->decodedValue() : $member;
             } catch (Throwable) {
                 return null;
             }
@@ -168,9 +168,9 @@ final class Dictionary implements StructuredFieldOrderedMap
     /**
      * Returns the Item value of a specific key if it exists and is valid otherwise returns null.
      *
-     * @return InnerList<int, Item>|ByteSequence|Token|float|int|bool|string|null
+     * @return InnerList<int, Item>|float|int|bool|string|null
      */
-    public function value(string|int $offset): InnerList|ByteSequence|Token|float|int|bool|string|null
+    public function value(string|int $offset): InnerList|float|int|bool|string|null
     {
         try {
             $member = $this->get($offset);
@@ -179,7 +179,7 @@ final class Dictionary implements StructuredFieldOrderedMap
         }
 
         if ($member instanceof Item) {
-            return $member->value;
+            return $member->decodedValue();
         }
 
         return $member;
@@ -275,7 +275,7 @@ final class Dictionary implements StructuredFieldOrderedMap
     private static function filterForbiddenState(InnerList|Item $member): InnerList|Item
     {
         foreach ($member->parameters as $item) {
-            if ($item->parameters->isNotEmpty()) {
+            if ($item->parameters->hasMembers()) {
                 throw new ForbiddenStateError('Parameters instances can not contain parameterized Items.');
             }
         }